refactor: dish 하위 패키지 dish 패키지로 통합
