+++ application/app-api/src/main/java/core/startup/mealtoktok/api/Dish/CreateDishApi.java
@@ -2,10 +2,9 @@ package core.startup.mealtoktok.api.dish;
 
 import core.startup.mealtoktok.api.dish.request.DishRequest;
 import core.startup.mealtoktok.common.dto.Response;
-import core.startup.mealtoktok.domain.DishStore.TargetDishStore;
+import core.startup.mealtoktok.domain.dish.TargetDishStore;
 import core.startup.mealtoktok.domain.dish.CreateDishService;
-import core.startup.mealtoktok.domain.dishCategory.TargetDishCategory;
-import core.startup.mealtoktok.domain.user.TargetUser;
+import core.startup.mealtoktok.domain.dish.TargetDishCategory;
 import core.startup.mealtoktok.domain.user.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
