+++ domain/src/main/java/core/startup/mealtoktok/domain/dish/DishAppender.java
@@ -1,7 +1,5 @@
 package core.startup.mealtoktok.domain.dish;
 
-import core.startup.mealtoktok.domain.DishStore.DishStore;
-import core.startup.mealtoktok.domain.dishCategory.DishCategory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
