+++ domain/src/main/java/core/startup/mealtoktok/domain/dish/DishCategoryReader.java
@@ -1,4 +1,4 @@
-package core.startup.mealtoktok.domain.dishCategory;
+package core.startup.mealtoktok.domain.dish;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
