+++ domain/src/main/java/core/startup/mealtoktok/domain/dish/DishRepository.java
@@ -1,8 +1,5 @@
 package core.startup.mealtoktok.domain.dish;
 
-import core.startup.mealtoktok.domain.DishStore.DishStore;
-import core.startup.mealtoktok.domain.dishCategory.DishCategory;
-
 import java.util.List;
 
 public interface DishRepository {
