+++ domain/src/main/java/core/startup/mealtoktok/domain/dish/DishStoreReader.java
@@ -1,4 +1,4 @@
-package core.startup.mealtoktok.domain.DishStore;
+package core.startup.mealtoktok.domain.dish;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
