+++ domain/src/main/java/core/startup/mealtoktok/domain/dish/OperatingHour.java
@@ -1,4 +1,4 @@
-package core.startup.mealtoktok.domain.DishStore;
+package core.startup.mealtoktok.domain.dish;
 
 import java.time.LocalTime;
