+++ domain/src/main/java/core/startup/mealtoktok/domain/dish/ReadDishService.java
@@ -1,11 +1,5 @@
 package core.startup.mealtoktok.domain.dish;
 
-import core.startup.mealtoktok.domain.DishStore.DishStore;
-import core.startup.mealtoktok.domain.DishStore.DishStoreReader;
-import core.startup.mealtoktok.domain.DishStore.TargetDishStore;
-import core.startup.mealtoktok.domain.dishCategory.DishCategory;
-import core.startup.mealtoktok.domain.dishCategory.DishCategoryReader;
-import core.startup.mealtoktok.domain.dishCategory.TargetDishCategory;
 import core.startup.mealtoktok.domain.user.TargetUser;
 import core.startup.mealtoktok.domain.user.UserReader;
 import lombok.RequiredArgsConstructor;
