+++ domain/src/main/java/core/startup/mealtoktok/domain/dish/TargetDishCategory.java
@@ -1,4 +1,4 @@
-package core.startup.mealtoktok.domain.dishCategory;
+package core.startup.mealtoktok.domain.dish;
 
 public record TargetDishCategory(
         Long categoryId
