+++ domain/src/main/java/core/startup/mealtoktok/domain/dish/TargetDishStore.java
@@ -1,4 +1,4 @@
-package core.startup.mealtoktok.domain.DishStore;
+package core.startup.mealtoktok.domain.dish;
 
 public record TargetDishStore(
         Long storeId
