+++ domain/src/main/java/core/startup/mealtoktok/domain/dish/UpdateDishService.java
@@ -1,8 +1,5 @@
 package core.startup.mealtoktok.domain.dish;
 
-import core.startup.mealtoktok.domain.DishStore.DishStore;
-import core.startup.mealtoktok.domain.DishStore.DishStoreReader;
-import core.startup.mealtoktok.domain.DishStore.TargetDishStore;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
