+++ domain/src/main/java/core/startup/mealtoktok/domain/dish/exception/DishErrorCode.java
@@ -10,7 +10,9 @@ import static core.startup.mealtoktok.common.consts.MealTokTokConstant.*;
 public enum DishErrorCode implements BaseErrorCode {
 
     DISH_NOT_FOUND(NOT_FOUND, "DISH_404_1", "반찬의 정보를 찾지 못했습니다."),
-    DISH_NAME_ALREADY_EXISTS(CONFLICT, "DISH_409_1","이미 존재하는 반찬 이름입니다.");
+    DISH_NAME_ALREADY_EXISTS(CONFLICT, "DISH_409_1","이미 존재하는 반찬 이름입니다."),
+    DISH_STORE_NOT_FOUND(NOT_FOUND, "DISH_STORE_404_1", "반찬 가게의 정보를 찾지 못했습니다."),
+    DISH_CATEGORY_NOT_FOUND(NOT_FOUND, "DISH_CATEGORY_404_1", "반찬 카테고리의 정보를 찾지 못했습니다.");
 
     private final Integer status;
     private final String errorCode;
