+++ infra/src/main/java/core/startup/mealtoktok/infra/dish/entity/DishEntity.java
@@ -2,8 +2,6 @@ package core.startup.mealtoktok.infra.dish.entity;
 
 import core.startup.mealtoktok.domain.dish.Dish;
 import core.startup.mealtoktok.domain.dish.DishInfo;
-import core.startup.mealtoktok.infra.dishCategory.entity.DishCategoryEntity;
-import core.startup.mealtoktok.infra.dishStore.entity.DishStoreEntity;
 import jakarta.persistence.*;
 import lombok.*;
