+++ infra/src/main/java/core/startup/mealtoktok/infra/dish/exception/DishStoreNotFoundException.java
@@ -1,14 +1,13 @@
-package core.startup.mealtoktok.infra.dishStore.exception;
+package core.startup.mealtoktok.infra.dish.exception;
 
 import core.startup.mealtoktok.common.exception.InfraException;
-import core.startup.mealtoktok.domain.DishStore.exception.DishStoreErrorCode;
-
+import core.startup.mealtoktok.domain.dish.exception.DishErrorCode;
 
 public class DishStoreNotFoundException extends InfraException {
 
     public static InfraException EXCEPTION = new DishStoreNotFoundException();
 
     private DishStoreNotFoundException() {
-        super(DishStoreErrorCode.DISH_STORE_NOT_FOUND);
+        super(DishErrorCode.DISH_STORE_NOT_FOUND);
     }
 }
