+++ infra/src/main/java/core/startup/mealtoktok/infra/dish/repository/JpaDishCategoryRepository.java
@@ -1,6 +1,6 @@
-package core.startup.mealtoktok.infra.dishCategory.repository;
+package core.startup.mealtoktok.infra.dish.repository;
 
-import core.startup.mealtoktok.infra.dishCategory.entity.DishCategoryEntity;
+import core.startup.mealtoktok.infra.dish.entity.DishCategoryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaDishCategoryRepository extends JpaRepository<DishCategoryEntity, Long> {
