+++ infra/src/main/java/core/startup/mealtoktok/infra/dish/repository/JpaDishStoreRepository.java
@@ -1,6 +1,6 @@
-package core.startup.mealtoktok.infra.dishStore.repository;
+package core.startup.mealtoktok.infra.dish.repository;
 
-import core.startup.mealtoktok.infra.dishStore.entity.DishStoreEntity;
+import core.startup.mealtoktok.infra.dish.entity.DishStoreEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaDishStoreRepository extends JpaRepository<DishStoreEntity, Long> {
